git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@259 d0c010d8-bb82-477e-a075-112140230451

+++ src/CWptDB.cpp
@@ -178,8 +178,8 @@ void CWptDB::saveGPX(CGpx& gpx)
         }
         QDomElement waypoint = gpx.createElement("wpt");
         root.appendChild(waypoint);
-        waypoint.setAttribute("lat",(*wpt)->lat);
-        waypoint.setAttribute("lon",(*wpt)->lon);
+        waypoint.setAttribute("lat",(double)(*wpt)->lat);
+        waypoint.setAttribute("lon",(double)(*wpt)->lon);
 
         if((*wpt)->ele != 1e25f){
             QDomElement ele = gpx.createElement("ele");
